Upgraded for Minecraft-1.7
Several things have changed with Minecraft-1.7, most importantly the
protocol. One of the more convenient additions to the protocol is the
sending of the client protocol version in the handshake request. This
enables the FakeServer to detect the version of the client, and conform
to the client during runtime.

The second major addition is the ability to supply the client with a
favicon. The icon is displayed next to the server listing in the server
list. By default, a warning symbol is sent to the client.

This patch removes support for clients prior to Minecraft-1.7.